!userinfo finished
This catches the !userinfo command up to the end of Threebow's Discord.js bot tutorial.

I had to do quite a bit oof troubleshooting to there the command to work. Its a good thing that I was given so many tools from troubleshooting code that isn't working.

The culprit was the field for the Full Username the addfield function didn't like the way the string was put together I guess.